example_usempi: expand the example
Mirror the structure of the C example: split the sends, recvs, and
allgathers into their own subroutines.  Also properly use MPI_Count
types and integers for invoking the BigCount variants of the
subroutines.

Fortran wins the "that's a really weird way of expressing a giant
integer" award of the day.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>