More sweeping changes: actually read from status
Update MPI_Get_elements[_x] to actually read from the status (and fill
in a bogus value on the status before reading it).  Note that MPI-3.1
says that if the size of the count argument is too small to hold the
actual count, we just fill in MPI_UNDEFINED.  That makes this whole
endeavour, while painstaking, rather straightforward.

Also used this opprotunity to consolidate "global" header files into
the include/ directory (vs. being scattered around multiple
directories).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>