Remove mpi.hpp (including all C++ overloading)
We're just going to have mpi.h, and there will be no C++ overloading
at all.  If you compile in C++, it's like you're compiling with a
compiler that simply does not support C11 _Generic (which is true!).

That being said, all the C symbols will still be prototyped in <mpi.h>
(e.g., MPI_Send_l() and friends) and available to link against.  You
just won't have the C11 _Generic auto-selectors to pick which flavor
to invoke.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>