Rename all functions from _l to _x
From 2019-07-23 Forum Virtual Meeting: there was a general preference
for the _x suffix over the _l suffix (i.e., to keep going with what we
started in MPI-3.1).

The one exception is the ~5 Fortran functions that already have
MPI_FOO_X interfaces in MPI-3.1:

1. We're not going to break backwards compatibility (e.g., by removing
   MPI_FOO_X in MPI-4.0)
2. As such, since an interface for MPI_FOO_X already exists, we can't
   make the MPI_Count-flavored procedure in the MPI_FOO interface have
   a _x_f08 suffix (because the MPI_FOO_X interface already has that
   suffix).
3. For these ~5 functions only (e.g., MPI_GET_ELEMENTS), they will use
   a _l_f08 suffix for the MPI_Count-flavored procedure in the MPI_FOO
   interface.  All other MPI_Count-flavored procedures in MPI_FOO
   interfaces will use a _x_f08 suffix.

Also expand the C example to explicitly call the _x functions (because
that was also a desire from the 2019-07-23 Forum Virtual Meeting).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>